Very smart telescope find files now, cmp abort change
Added falling back to find_files if git_files can't find a .git directory, so that the find files thingy always sees dotfiles.
Changed the abort key for completions in insert mode to Esc: hit ESC 2
times to get out of completion mode and then out of insert to normal.